+++ preferences/src/main.rs
@@ -12,8 +12,11 @@ slint::include_modules!();
 
 fn main() -> anyhow::Result<()> {
     win_dbg_logger::init();
-    slint::BackendSelector::new()
-        .with_winit_window_attributes_hook(|attrs| attrs.with_transparent(false))
+    let mut sel = slint::BackendSelector::new();
+    if is_vm() {
+        sel = sel.renderer_name("software".to_string());
+    }
+    sel.with_winit_window_attributes_hook(|attrs| attrs.with_transparent(false))
         .select()?;
     if env::args()
         .any(|arg| arg == "/about" || arg == "--about" || arg == "chewing-preferences://about/")
@@ -23,4 +26,12 @@ fn main() -> anyhow::Result<()> {
         config::run()?;
     }
     Ok(())
+}
+
+fn is_vm() -> bool {
+    if let Some(mb) = sysinfo::Motherboard::new() {
+        let name = mb.name().unwrap_or_default();
+        return ["Virtual Machine"].contains(&name.as_str());
+    }
+    true
 }