+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Add missing bopomofo mapping to the GinYieh layout
 - Correct IBM layout mapping
+- Fallback to software renderer on Hyper-V
 
 ### 🚜 Refactor